refactor: move generate metadat to page
